requirements.txt
Django==1.11
django-extensions==1.9.8
dj-database-url==0.4.2
gunicorn==19.7.1
pkg-resources==0.0.0
psycopg2==2.7.3.2
PyJWT==1.5.3
pytz==2017.3
six==1.11.0
typing==3.6.2
